most of paco errors fixed

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: beldemir <beldemir@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/27 19:05:15 by beldemir          #+#    #+#             */
-/*   Updated: 2024/10/28 05:45:13 by beldemir         ###   ########.fr       */
+/*   Updated: 2024/10/28 08:45:17 by beldemir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ int	ft_print_s(char *s);
 int	ft_print_p(void *p);
 int	ft_print_i_d(int i, int flag);
 int	ft_print_u(unsigned int n);
-int	ft_print_x(unsigned long n, const char c, int flag);
+int	ft_print_x(unsigned int n, const char c, int flag);
 
 int	ft_check_flag(const char c);
 int	ft_check_parameter(const char c, va_list args, int flag);
